src/other/other.sql/requirements.txt
oracledb==2.2.1
pyodbc==5.0.1
mysql-connector-python==8.4.0
pymssql==2.3.0
psycopg2-binary==2.9.9
docker==7.1.0